backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
rank-bm25
InstructorEmbedding
Tashaphyne
